+++ lib/utils/constants/drawer_constants.dart
@@ -11,8 +11,10 @@ abstract class DrawerConstants {
   static const NOTICE_BOARD = 7;
   static const TIMETABLE = 8;
   static const SCRUTINI = 9;
+  static const STATS = 10;
+
   // static const ANNO_PRECEDENTE = 10;
   // static const WEB = 11;
-  static const SETTINGS = 10;
-  static const SHARE = 11;
+  static const SETTINGS = 11;
+  static const SHARE = 12;
 }